requirements.txt
Django==3.1.3
djangorestframework==3.12.2
pyyaml==5.3.1
django-polymorphic==3.0.0
lorem==0.1.1
django-rest-polymorphic==0.1.9
celery==5.0.2
psycopg2-binary==2.8.6
django-celery-results==2.0.0
